use float for mint fee

+++ templates/digital-asset-template/frontend/pages/CreateCollection.tsx
@@ -313,7 +313,7 @@ export function CreateCollection() {
               type="number"
               value={mintFeePerNFT}
               onChange={(e) => {
-                setMintFeePerNFT(parseInt(e.target.value));
+                setMintFeePerNFT(parseFloat(e.target.value));
               }}
             />
           </div>
